requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pymupdf
google-cloud-vision
google-api-python-client
google-auth-oauthlib
google-auth
pytest
requests
boto3